Extract household inspection into trait
